Factor out 3rd party dependencies from child to parent POM
Dependency and plugin version detection can be performed in a dependencyManagement and
pluginManagement block instead of individually in the dependencies and plugins block. These settings
are also inherited by the POM files. By moving the version definitions to the main POM, the
management becomes easier and version clashes are avoided.