Remove extensions section in applet pom.xml
The extensions element is not supported by the maven modifications of IntelliJ Idea. By moving the element to an
extensions entry in the plugin, the bug in IntelliJ is circumvented.
Without the toplevel extensions element, the pack200 plugin must be defined in the plugins section rather than in the
pluginManagement section.