Fallback to TLS 1.0 when server implementations are errornous (fixes #177)
If grabbing the TCToken using TLSv1.1 results in an error in the TLSProtocolHandler we now try
again using TLSv1.0. This also applies to PAOS connections.

The rationale behind this is that most production eID servers are not capable of handling TLS 1.1
connections as required by TR-03130 (Version 1.1.2) Section 2. We are currently speaking with the
eID Server producers to address this problem, so that we can remove the fallback.