Perform PAOS in background task and redirect browser immediately if activated by Object
eID Servers using the old activation method rely on the browser coming to the refresh address
directly after triggering the app. If that is not the case, the process does not terminate and a
timeout is thrown by the eID server.

Disable SNI for alle eID servers

Because of the handcrafted TLS implementations in the eID servers, SNI is often not supported
correctly. It is also the case that these servers do not need SNI, because they usually have a
dedicated IP for the name.