Add code that simulates ld under linux and use it to find pcsclite (fixes #216)
ld uses LD_LIBRARY_PATH, ld.so.cache and fixed base paths to find the library. The SmartcardIO PCSC
implementation only uses the fixed base paths and Java specific load paths. The LinuxLibraryFinder
class finds a library according to the mechanism described in the ld.so(8) man page. In case no
library could be found, an error is raised instead of relying on the SmartcardIO to find the lib
nevertheless.