Fix start problems of the Android App
Under certain circumstances the App will hang on startup because the is libusb waiting for the
device descriptors and therefore blocking the pcsc daemon and the pcsc factory what finally blocks
the whole initialization of the client.
To prevent this we enforce the DeviceOpenActivity to be always the first activity that will be
started. It will provide the device descriptors needed by libusb.
Secondly there was an error in the marshaller that disables the marshalling of card info files.
Finally waiting for thread pools on shutdown always endend up waiting the whole waiting delay
and then forcing the shutdown. Therefore we can immediately force the shutdown.

Conflicts:
	clients/android/AndroidManifest.xml
	clients/android/src/main/java/org/openecard/android/activities/IntentHandlerActivity.java

Conflicts:
	clients/android/AndroidManifest.xml
	clients/android/src/main/java/org/openecard/android/ApplicationContext.java
	clients/android/src/main/java/org/openecard/android/activities/IntentHandlerActivity.java
	wsdef/android-marshaller/src/main/java/org/openecard/ws/android/AndroidMarshaller.java