Wait for PAOSTask at the end of the activation handler before returning to the transport
On the android client, the app shuts down after receiving the TCTokenResponse, which is wrong when
the PAOS task is executed in the background. The problem is solved, when the handler waits for the
task to finish. There is no need to perform transport specific actions to wait for the task.