Make return-to-websession more TR-03112 conforming
The algorithm in sec. 3.4.5 states, that if the refresh address conforms to the
Same-Origin-Policy (SOP), then no GET on the adress is performed. In case The refresh is not SOP,
then it must be followed and only stopped, when the SOP endpoint returns a redirect which in turn is
not retrieved by the client. This algorithm is sick ...