Update the code according to a fix in  org.openecard.common.sal.anytype.CryptoMarkerType
This commit fixes issue 298.
The CryptoMarkerType stores now a list of certificates instead of a single certificate. Per convention the 
first certificate in the list is now the certificate to use for TLS authentication or signature creation or ...
The classes depending on the CryptoMarkerType are updated to use this first certificate.