Add the AddonManager to the Constructor of the file registry.
Splits also the loadLoadOnStartAddons into two methods to load all addons and to load a single addon.
This is required for the case a new addon was added and the description contains a LoadOnStartup in
the manifest. In this case the manager have to invoke the init() function of the specific action.