Remove addon correctly and put created addons in a cache
The behavior before this commit was to create a new instance of specific class of a addon on every 
get.*[Protocol|Action] call. This had the effect that the destroy method wasn't callable because the 
instances were not stored. This commit solved this problem by adding data structures which cache these 
protocols and actions. Furthermore the unload and shutdown methods are added to AddonManager to 
correctly destroy a addon on request or shutdown. The implementation of the removal of addons required 
the extension of the FileRegistry which now associate the AddonSpecifiaction with the file name.