Fix the creation of the DefaultSettingsGroup.
The previous implementation ignored configurations of the actions and protocols and 
returned always a Settings group which contained just the general Configuration of an 
add-on. This commit fixes this behavior by modifying the constructor which takes now 
a list of configuration entries. The SetttingsGroup is now created from this list instead of 
the AddonSpecification.