Extend API of Body class for data exchange with bindings
The old Body class has been based on a DOM node, which is not capable of representing 
anything else than XML data. The new API is String based and has an additional flag to 
mark the data as Base64 encoded.
Furthermore the classes received documentation and various set and get functions.