fix: Disable UI theme init to fix hang when calling gtk
Gtk calls are all over the place in swing and the dorkbox systray. When overwriting the theme, there seems to be some
interference with Gtk, leading to hangs in either swing or systray, depending on whether the tray was initialized swing.

Not setting the theme seems to fix the app on the affected platforms.