feat: Remove false from ACL
We require that an ACL contains a branch which resolves to true. We further define the empty tree as being false. By
doing that there is no need to explicitly express false in the ACL. Removing false also makes the algorithms a little
bit easier, as false is simply the empty list in a branch.