feat: Implement PIN DID
This includes changes in the CIF too, as the current definition was not sufficient to decide how verify, modify, and
reset commands need to be constructed.

Working on this made it obvious, that the implementation of the hardware modify command is wrong for anything more
complex. The implementation has been written for the nPA where modify is called only with the new PIN. Cases with old
and new reference data or two different encoded PINs are currently not possible. This has to be addressed in the future.