feat: Handle additional certificates from EAC2Input correctly
The CAR values are only sent when no chain can be built. In that case the eID-Server will return additional certificates
if possible. These certificates can be used to build the CVC chain that the card understands.

This patch handles the case that no chain can be built as a client error. In fact it is undecidable, because it depends
on the intention of the user. If a test card is used with a production server, whose fault is it. Because the user
ultimately selects the card, we handle it as a client error.